UnitTests - Switch ext tests from v3 to v4 test trait
